Create internal/time
