+++ internal/time/time.go
@@ -19,19 +19,16 @@ along with this program.  If not, see http://www.gnu.org/licenses/.
 	@License:		https://github.com/edoardottt/boggart/blob/main/LICENSE
 */
 
-package api
+package time
 
-import "errors"
+import "time"
 
-var (
-	ErrPossibleTopValue      = errors.New("possible values for top: method / path / body")
-	ErrNoStatsIP             = errors.New("no stats available for the specified IP")
-	ErrIDDefined             = errors.New("if id is defined, no other parameters need to be defined")
-	ErrDateDefined           = errors.New("if date is defined, lt and gt must be blank")
-	ErrLtBeforeGt            = errors.New("lt cannot be before gt")
-	ErrHTTPMethodUnknown     = errors.New("http method unknown")
-	ErrInvalidIP             = errors.New("ip address is not valid")
-	ErrStringToIntConversion = errors.New("failed to convert string to int")
-	ErrRetrievingData        = errors.New("error while retrieving data")
-	ErrIntegerRange          = errors.New("integer not in the range")
-)
+// TranslateTime.
+func TranslateTime(input string) (time.Time, error) {
+	t, err := time.Parse("2006-01-02T15:04:05-0700", input)
+	if err != nil {
+		return time.Time{}, ErrDatetimeFormat
+	}
+
+	return t, nil
+}
